+++ pysat/instruments/methods/testing.py
@@ -4,9 +4,12 @@ import datetime as dt
 import numpy as np
 import os
 import pandas as pds
+import time
 import warnings
 
 import pysat
+from pysat.utils import NetworkLock
+from pysat.utils import time as putime
 
 logger = pysat.logger
 
@@ -515,3 +518,96 @@ def define_range():
                  'angle': [0.0, 2.0 * np.pi]}
 
     return def_range
+
+
+def create_files(inst, start, stop, freq='1D', use_doy=True,
+                 root_fname='pysat_testing_{year:04d}_{day:03d}.txt',
+                 version=False, content=None, timeout=None):
+    """Create a file set using the year and day of year.
+
+    Parameters
+    ----------
+    inst : pysat.Instrument
+        A test instrument, used to generate file path
+    start : dt.datetime
+        The date for the first file to create
+    stop : dt.datetime
+        The date for the last file to create
+    freq : str
+        Frequency of file output.  Codes correspond to pandas.date_range
+        codes (default='1D')
+    use_doy : bool
+        If True use Day of Year (doy), if False use day of month and month.
+        (default=True)
+    root_fname : str
+        The format of the file name to create. Supports standard pysat template
+        variables 'year', 'month', 'day', 'hour', 'minute', 'second', 'version',
+        'revision', 'cycle'. (default='pysat_testing_{year:04d}_{day:03d}.txt')
+    version : bool
+        If True, iterate over version / revision / cycle. If False,
+        ignore version / revision / cycle. (default=False)
+    content : str
+        Custom text to write to temporary files (default=None)
+    timeout : float
+        Time is seconds to lock the files being created.  If None, no timeout is
+        used.  (default=None)
+
+    Examples
+    --------
+    >>> import datetime as dt
+    >>> inst = pysat.Instrument('pysat', 'testing')
+    >>> root_fname='pysat_testing_{year:04d}_{month:02d}_{day:02d}.txt'
+    >>> create_files(inst, dt.datetime(2008, 1, 1), dt.datetime(2008, 12, 31),
+    ...              root_fname=root_fname, use_doy=False)
+
+
+    Command creates empty files located at `inst.files.data_path`, one per day,
+    spanning 2008, where `year`, `month`, and `day` are filled in using the
+    provided template string appropriately. Produced files look like:
+    'pysat_testing_2008_01_01.txt'
+
+    >>> create_files(inst, dt.datetime(2008, 1, 1), dt.datetime(2008, 12, 31))
+
+    In this case we use all of the function defaults which produces a set of
+    files on a daily basis, labeled by year and day of year. Produced files look
+    like: 'pysat_testing_2008_001.txt'
+
+    """
+
+    # Define the time range and file naming variables
+    dates = putime.create_date_range(start, stop, freq=freq)
+
+    if version:
+        versions = np.array([1, 2])
+        revisions = np.array([0, 1])
+        cycles = np.array([0, 1])
+    else:
+        versions = [None]
+        revisions = [None]
+        cycles = [None]
+
+    # Create empty files
+    for date in dates:
+        yr, doy = putime.getyrdoy(date)
+        if not use_doy:
+            doy = date.day
+        for version in versions:
+            for revision in revisions:
+                for cycle in cycles:
+
+                    fname = os.path.join(inst.files.data_path,
+                                         root_fname.format(year=yr,
+                                                           day=doy,
+                                                           month=date.month,
+                                                           hour=date.hour,
+                                                           minute=date.minute,
+                                                           second=date.second,
+                                                           version=version,
+                                                           revision=revision,
+                                                           cycle=cycle))
+                    with NetworkLock(fname, 'w') as fout:
+                        if content is not None:
+                            fout.write(content)
+                        if timeout is not None:
+                            time.sleep(timeout)
+    return